chore(test): restructure tests
