+++ packages/audioplayers/example/integration_test/tabs/context_tab_test.dart
@@ -1,0 +1,15 @@
+import 'package:flutter_test/flutter_test.dart';
+
+import '../platform_features.dart';
+import '../source_test_data.dart';
+
+Future<void> testContextTab(
+  WidgetTester tester,
+  SourceTestData audioSourceTestData,
+  PlatformFeatures features,
+) async {
+  // Audio context
+  // TODO(Gustl22): test generic flags
+  // await tester.tap(find.byKey(const Key('audioContextTab')));
+  // await tester.pumpAndSettle();
+}
