+++ packages/audioplayers/example/integration_test/tabs/logs_tab_test.dart
@@ -1,0 +1,14 @@
+import 'package:flutter_test/flutter_test.dart';
+
+import '../platform_features.dart';
+import '../source_test_data.dart';
+
+Future<void> testLogsTab(
+  WidgetTester tester,
+  SourceTestData audioSourceTestData,
+  PlatformFeatures features,
+) async {
+  // TODO(Gustl22): may test logs
+  // await tester.tap(find.byKey(const Key('loggerTab')));
+  // await tester.pumpAndSettle();
+}
